Added delete mgs conditions.

+++ src/handlers/leave/index.js
@@ -3,8 +3,15 @@ const loadHandler = require('../load')
 
 module.exports = () => [
   async (ctx) => {
-    await ctx.deleteMessage(ctx.session.board)
-    await ctx.deleteMessage(ctx.session.actions)
+    if (ctx.session.board) {
+      await ctx.deleteMessage(ctx.session.board)
+      ctx.session.board = null
+    }
+
+    if (ctx.session.actions) {
+      await ctx.deleteMessage(ctx.session.actions)
+      ctx.session.actions = null
+    }
 
     await loadHandler()[0](ctx)
